refactor: PathVariable 을 `postId`에서 `id`로 변경
