+++ src/pages/human-rights/[id]/mockQueries.ts
